UNTESTED: Fixed broken compilation on macOS.

+++ include/gui.h
@@ -29,10 +29,10 @@ enum ElementTypes {
 class GUI {
     public:
         GUI();
-        void Button(int ID, int PosX, int PosY, string ButtonText, int Width, int Height, Color ButtonColor, Font ButtonFont, int FontSize, Color TextColor, void (*onclick)()=NULL);
-        void TextBox(int ID, int PosX, int PosY, string PlaceholderText, int Width, int Height, Color BoxColor, Font TextFont, int FontSize, Color TextColor, bool ShowLabel=false, string LabelText="", int LabelPadding=120);
-        void Grid(double PosX, double PosY, double Width, double Height, double Spacing, double Radius, Color GridColor, bool GenerateLayoutFromGrid=false);
-        void Window(int ID, double PosX, double PosY, double Width, double Height, string Title="", Font TitleFont=GetFontDefault(), float FontSize=16);
+        void Button(int ID, float PosX, float PosY, string ButtonText, float Width, float Height, Color ButtonColor, Font ButtonFont, int FontSize, Color TextColor, void (*onclick)()=NULL);
+        void TextBox(int ID, float PosX, float PosY, string PlaceholderText, float Width, float Height, Color BoxColor, Font TextFont, int FontSize, Color TextColor, bool ShowLabel=false, string LabelText="", int LabelPadding=120);
+        void Grid(float PosX, float PosY, float Width, float Height, float Spacing, double Radius, Color GridColor, bool GenerateLayoutFromGrid=false);
+        void Window(int ID, float PosX, float PosY, float Width, float Height, string Title="", Font TitleFont=GetFontDefault(), float FontSize=16);
         void WindowFromGrid(int ID, int X1, int Y1, int X2, int Y2, string Title="", Font TitleFont=GetFontDefault(), float fontSize=16);
         Vector2 FromGridLayout(int x, int y);
         void MainEventLoop(string DPATH);
@@ -69,11 +69,11 @@ class GUI {
 
         class GridLayout {
             public:
-                double PosX;
-                double PosY;
-                double Width;
-                double Height;
-                double Spacing;
+                float PosX;
+                float PosY;
+                float Width;
+                float Height;
+                float Spacing;
         };
         GridLayout CurrentGridLayout;
 
@@ -86,8 +86,8 @@ class GUI {
         int MouseY;
         double InitialMouseX = 0;
         double InitialMouseY = 0;
-        int rX;
-        int rY;
+        float rX;
+        float rY;
         int WinMoved = -1;
         bool ShowGrid;
         bool ValidWindowPosition = false;
